Fix SQL migration ROUND function type error
- Cast the division result to NUMERIC before using ROUND
- Fixed in both schema.sql and migration file
- Added .env file with Supabase credentials

+++ database/migrations/001_initial_schema.sql
@@ -123,7 +123,7 @@ SELECT
     ba.total_parcels,
     ba.vacant_parcels,
     ba.vacant_buildings,
-    ROUND((ba.vacant_parcels::FLOAT / NULLIF(ba.total_parcels, 0) * 100), 2) as vacancy_rate,
+    ROUND((ba.vacant_parcels::FLOAT / NULLIF(ba.total_parcels, 0) * 100)::NUMERIC, 2) as vacancy_rate,
     ba.avg_assessed_value,
     ba.recent_sales_count,
     ba.tax_delinquent_percentage,
